moved ThrowAwayIndexLoader to braindecode.util for import in classifier.py and regressor.py
